Move protectedResource to shared package
